% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybayes.R
\name{mybayes}
\alias{mybayes}
\title{Bayes Theorem Function}
\usage{
mybayes(x, y, z)
}
\arguments{
\item{x}{corresponds to P(A) of theorem}

\item{y}{corresponds to P(B) of theorem}

\item{z}{corresponds to P(B|A) of theorem}
}
\value{
the probability value of an event using Bayes Theorem (i.e. P(A|B))
}
\description{
Bayes Theorem Function
}
\examples{
mybayes(0.90, 0.10, 0.5)

}
