% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Simulation}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of Iterations}

\item{n}{Sample Size}

\item{p}{Probability of Success}
}
\value{
A table and barplot
}
\description{
Binomial Simulation
}
\examples{
mybin()

}
