% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of interations}

\item{x}{vector}

\item{fun}{which type of function is used in apply() i.e. mean, median, etc.}

\item{alpha}{confindence percentage in decimal form}

\item{...}{any additional arguments}
}
\value{
a plot and a list of values if needed (see return function at bottom of function)
}
\description{
Title
}
\examples{
myboot2(x=c(3,2,4,5), iter = 10000, fun="mean", alpha = 0.05)

}
