% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem Function for Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{number of trials}

\item{iter}{iterations}

\item{lambda}{rate}

\item{...}{}
}
\value{
a series of plots
}
\description{
Central Limit Theorem Function for Poisson
}
\examples{
mycltp(n=3,iter=10000,lambda=4)

}
