% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Plotting and Calculation}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{is the mean}

\item{sigma}{is the standard deviation}

\item{a}{is our supplied x value}
}
\value{
a plot of the curve, shaded area, and area probability to the console
}
\description{
Normal Curve Plotting and Calculation
}
\examples{
myncurve(mu=10,sigma=5, a=6)

}
