% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Title}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{x0 value; such as 1}

\item{delta}{delta; value such as 0.000001}

\item{llik}{used to pass a function; see example for more information (each dpois has '*' between them but it broke this documentation so they were removed. Please include those is using the example :) )}

\item{xrange}{the numerical range of x}

\item{parameter}{parameter; such as 'lambda'}
}
\value{
a plot and all the dat regarding the plot
}
\description{
Title
}
\examples{
myNRML(x0=1,delta=0.000001,llik=function(x) log(dpois(4,x)dpois(6,x)dpois(7,x)dpois(6,x)dpois(5,x)),xrange=c(0,20),parameter="lambda" )

}
