% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_lm.R
\name{piecewise_lm}
\alias{piecewise_lm}
\title{Piecewise Function}
\usage{
piecewise_lm(datadf, xk)
}
\arguments{
\item{datadf}{the data from the correct file}

\item{xk}{the variable used in piecewise linear model}
}
\value{
a plot and various values of data
}
\description{
Piecewise Function
}
\examples{
piecewise_lm(spruce.df, 18)

}
