% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.df.R
\docType{data}
\name{spruce.df}
\alias{spruce.df}
\title{Spruce data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{Height}{height}
\item{BHDiameter}{bhdiameter}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
spruce.df
}
\description{
A data set called SPRUCE.csv
}
\keyword{datasets}
